+++ src/cli/matches.rs
@@ -1,14 +1,16 @@
-use std::collections::HashMap;
 // NATIVE
 use std::thread;
 use std::time::Duration;
+use std::collections::HashMap;
 
 // by MYWAY
 use crate::cli::commands::Commands;
 use crate::log::log::{ Log, LogF };
-use crate::core::project::{ GraveyardList, Project, ProjectList, view_mission };
+
+use crate::core::project::{ Project, GenericList, User };
 use crate::core::filemanager::Fiman;
 use crate::core::errors::MyWayError;
+use crate::core::filemanager::ReturnReadType;
 
 // CRATES
 use chrono::{ Local };
@@ -17,6 +19,8 @@ use tequel_rs::rng::TequelRng;
 
 use semver::Version;
 use reqwest::header::USER_AGENT;
+
+
 
 pub fn check_for_updates(log: &Log) -> Option<String> {
     let current_version = env!("CARGO_PKG_VERSION");
@@ -38,11 +42,11 @@ pub fn check_for_updates(log: &Log) -> Option<String> {
         let latest = Version::parse(latest_version_str).ok()?;
 
         if latest > current {
-            log.hey_mw(&format!("Hey! This MyWay is deprecated, update now!"));
+            log.hey_mw(&format!("Hey! This MyWay is from older version, update now!"));
             log.hey_mw(&format!("{} -> {}", current, latest));
-            log.hey_mw("Use: cargo install myway-cli");
+            log.hey_mw(&format!("Use: {}", "cargo install myway-cli".yellow()));
         } else {
-            log.hey_mw(&format!("MyWay is updated!"));
+            log.hey_mw(&format!("MyWay is {}!", "updated".yellow()));
         }
     }
 
@@ -52,7 +56,7 @@ pub fn check_for_updates(log: &Log) -> Option<String> {
 
 
 
-pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut ProjectList, graveyard: &mut GraveyardList) -> Result<(), MyWayError> {
+pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut GenericList, graveyard: &mut GenericList, userdata: &mut User) -> Result<(), MyWayError> {
 
     let teq_rng: TequelRng = TequelRng::new();
 
@@ -109,8 +113,8 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
             let proj_description: String = log.quest_mandatory("description", "It's my project");
 
-            if proj_description.len() > 35 {
-                return Err(MyWayError::StringLengthLimitExceeded("description: <=35 8limit length".to_string()))
+            if proj_description.len() > 80 {
+                return Err(MyWayError::StringLengthLimitExceeded("description: <=80 limit length".to_string()))
             }
             
             log.hey("");
@@ -167,12 +171,17 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                 
                 time_created: format_data.to_string(),
 
+                trace: vec![],
+
                 is_finish: false
             };
 
             data.push(my_project);
-            files.write(&data, &files.mw_path.clone())?;
-
+            userdata.projects_created += 1;
+            
+            files.write(&ReturnReadType::User(userdata.clone()), &files.user_path.clone())?;
+            files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
+            
             Ok(())
 
         }
@@ -186,16 +195,13 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
             log.hey("");
             log.hey(&format!("{}", ">-----------------------------".dimmed()));
 
-            let mut target_uuid = None;
-
-            if let Some(uuid) = uuid {
-                target_uuid = data.iter().find(|p| p.uuid == *uuid).map(|p| p.uuid.clone());
-            }
-
-            if let Some(name) = name {
-                target_uuid = data.iter().find(|p| p.name == *name).map(|p| p.uuid.clone());
-            }
-
+            let target_uuid = if let Some(u) = uuid {
+                data.iter().find(|p| p.uuid == *u).map(|p| p.uuid.clone())
+            } else if let Some(n) = name {
+                data.iter().find(|p| p.name == *n).map(|p| p.uuid.clone())
+            } else {
+                None
+            };
 
             if let Some(uuid) = target_uuid {
 
@@ -206,8 +212,8 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                     
                     let mut proj_name = log.quest_mandatory("project_name", &proj.name);
     
-                    if proj_name.len() > 12 {
-                        return Err(MyWayError::StringLengthLimitExceeded("project_name: <=12 limit length".to_string()))
+                    if proj_name.len() > 20 {
+                        return Err(MyWayError::StringLengthLimitExceeded("project_name: <=20 limit length".to_string()))
                     }
 
                     if proj_name.contains(char::is_whitespace) {
@@ -224,8 +230,8 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
     
                     let proj_description: String = log.quest_mandatory("description", &proj.description);
     
-                    if proj_description.len() > 35 {
-                        return Err(MyWayError::StringLengthLimitExceeded("description: <=35 limit length".to_string()))
+                    if proj_description.len() > 80 {
+                        return Err(MyWayError::StringLengthLimitExceeded("description: <=80 limit length".to_string()))
                     }
 
                     log.hey("");
@@ -245,7 +251,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                     
                     log.hey_mw(&format!("{}", "Editted with successfully!".yellow()));
                     let creating_msg = format!("{}", "Updating your WAY... ".dimmed());
-                    let final_msg = format!("Now {} is updated on your WAY!", proj_name);
+                    let final_msg = format!("Now {} is updated on your WAY!", proj_name.yellow().bold());
                     
                     log.hey_mw(&creating_msg);
                     thread::sleep(Duration::from_millis(500));
@@ -266,6 +272,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                         status: proj.status.clone(),
                         
                         time_created: proj.time_created.clone(),
+                        trace: vec![],
     
                         is_finish: proj.is_finish.clone()
                     };
@@ -274,7 +281,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                         *proj = my_project;
                     }
     
-                    files.write(&data, &files.mw_path.clone())?;
+                    files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
 
                     Ok(())
 
@@ -296,13 +303,15 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
         // List all projects on way
         Commands::Way { oneline, complex, uuid, name, finish, working, status } => {
             
-            log.hey_mw(&format!("{}", "Walking on WAY..."));
+            log.hey_mw(&format!("{}", "Walking on Way!"));
             log.hey("");
-
+            log.hey(&format!("{}", ">-----------------------------".dimmed()));
+    
             if let Some(id) = uuid {
 
                 match data.iter().find(|p| p.uuid == *id) {
                     Some(project) => {
+                        println!("");
 
                         let is_finished = if project.is_finish {
                             "F"
@@ -310,55 +319,16 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                             "W"
                         };
 
-                        log.hey_mw(&format!("({}) {} {} {}", is_finished.bold().yellow(), project.uuid.italic(), project.name.bold(), format!("[{}]", project.versions.last().expect("")).italic()));
+                        log.hey_mw(&format!("({} : {}) {} {} {}", is_finished.bold().yellow(), project.status.yellow().bold(), project.uuid.italic(), project.name.bold(), format!("[{}]", project.versions.last().expect("")).italic()));
 
                         if !*oneline {
-                            if *complex {
-
-                                println!("\t{}", project.description.italic());
-                                println!("\tStatus: {}", project.status.bold());
-                                println!("\tMission: \"{}\"", view_mission(&project.mission).italic().yellow());
-                                println!("\tStack(s):");
-                                if project.stack.len() != 0 {
-                                    for i in project.stack.iter() {
-                                        println!("\t- {} ", i.bold())
-                                    }
-                                } else {
-                                    println!("\t {}", "No Stack(s)".bold())
-                                }
-                                println!("\tCreated at: {}", project.time_created.italic());
-
-                            } else {
-                                
-                                println!("\t{}", project.description.italic());
-                                println!("\tStatus: {}", project.status.bold());
-                                println!("\tMission: \"{}\"", view_mission(&project.mission).italic().yellow());
-                                println!("\tStack(s):");
-                                if project.stack.len() != 0 {
-                                    for i in project.stack.iter() {
-                                        println!("\t- {} ", i.bold())
-                                    }
-                                } else {
-                                    println!("\t {}", "No Stack(s)".bold())
-                                }
-
-                            }
+                            if *complex { log.hey_project(project, true); } 
+                            else { log.hey_project(project, false); }
                         } else {
-                            if *complex {
-                                println!("\t{}", project.description.italic());
-                                println!("\tStatus: {}", project.status.bold());
-                                println!("\tMission: \"{}\"", view_mission(&project.mission).italic().yellow());
-                                println!("\tStack(s):");
-                                if project.stack.len() != 0 {
-                                    for i in project.stack.iter() {
-                                        println!("\t- {} ", i.bold())
-                                    }
-                                } else {
-                                    println!("\t {}", "No Stack(s)".bold())
-                                }
-                                println!("\tCreated at: {}", project.time_created.italic());
-                            }
+                            if *complex { log.hey_project(project, true); }
                         }
+
+                        println!("");
 
                         Ok(())
                     }
@@ -371,64 +341,27 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
                 match data.iter().find(|n| n.name == *name) {
                     Some(project) => {
-                        
+                        println!("");
+
                         let is_finished = if project.is_finish {
                             "F"
                         } else {
                             "W"
                         };
 
-                        log.hey_mw(&format!("({}) {} {} {}", is_finished.bold().yellow(), project.uuid.italic(), project.name.bold(), format!("[{}]", project.versions.last().expect("")).italic()));
+                        log.hey_mw(&format!("({} : {}) {} {} {}", is_finished.bold().yellow(), project.status.yellow().bold(), project.uuid.italic(), project.name.bold(), format!("[{}]", project.versions.last().expect("")).italic()));
 
                         if !*oneline {
-                            if *complex {
-
-                                println!("\t{}", project.description.italic());
-                                println!("\tStatus: {}", project.status.bold());
-                                println!("\tMission: \"{}\"", view_mission(&project.mission).italic().yellow());
-                                println!("\tStack(s):");
-                                if project.stack.len() != 0 {
-                                    for i in project.stack.iter() {
-                                        println!("\t- {} ", i.bold())
-                                    }
-                                } else {
-                                    println!("\t {}", "No Stack(s)".bold())
-                                }
-                                println!("\tCreated at: {}", project.time_created.italic());
-
-                            } else {
-                                
-                                println!("\t{}", project.description.italic());
-                                println!("\tStatus: {}", project.status.bold());
-                                println!("\tMission: \"{}\"", view_mission(&project.mission).italic().yellow());
-                                println!("\tStack(s):");
-                                if project.stack.len() != 0 {
-                                    for i in project.stack.iter() {
-                                        println!("\t- {} ", i.bold())
-                                    }
-                                } else {
-                                    println!("\t {}", "No Stack(s)".bold())
-                                }
-
-                            }
+                            if *complex { log.hey_project(project, true); } 
+                            else { log.hey_project(project, false); }
                         } else {
-                            if *complex {
-                                println!("\t{}", project.description.italic());
-                                println!("\tStatus: {}", project.status.bold());
-                                println!("\tMission: \"{}\"", view_mission(&project.mission).italic().yellow());
-                                println!("\tStack(s):");
-                                if project.stack.len() != 0 {
-                                    for i in project.stack.iter() {
-                                        println!("\t- {} ", i.bold())
-                                    }
-                                } else {
-                                    println!("\t {}", "No Stack(s)".bold())
-                                }
-                                println!("\tCreated at: {}", project.time_created.italic());
-                            }
+                            if *complex { log.hey_project(project, true); }
                         }
 
+                        println!("");
+                        
                         Ok(())
+
 
                     }
                     None => {
@@ -458,6 +391,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
                 log.hey_mw(&format!("{} project(s) found", project_way.len()));
                 log.hey(&format!("{}", ">-----------------------------".dimmed()));
+                log.hey("");
 
                 for project in project_way.iter() {
 
@@ -493,16 +427,15 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
             log.hey_mw(&format!("{}", "On Way!"));
             log.hey("");
+            log.hey(&format!("{}", ">-----------------------------".dimmed()));
 
-            let mut target_uuid = None;
-
-            if let Some(uuid) = uuid {
-                target_uuid = data.iter().find(|p| p.uuid == *uuid).map(|p| p.uuid.clone());
-            }
-
-            if let Some(name) = name {
-                target_uuid = data.iter().find(|p| p.name == *name).map(|p| p.uuid.clone());
-            }
+            let target_uuid = if let Some(u) = uuid {
+                data.iter().find(|p| p.uuid == *u).map(|p| p.uuid.clone())
+            } else if let Some(n) = name {
+                data.iter().find(|p| p.name == *n).map(|p| p.uuid.clone())
+            } else {
+                None
+            };
 
             if let Some(uuid) = target_uuid {
 
@@ -531,8 +464,13 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                         log.hey_mw(&format!("{} was removed!", project.name.yellow()));
                         
                         data.retain(|p| p.uuid != uuid_removed);
+
+                        
+                        userdata.projects_giveup += 1;
+
+                        files.write(&ReturnReadType::User(userdata.clone()), &files.user_path.clone())?;
     
-                        files.write(&data, &files.mw_path.clone())?;
+                        files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
 
                         Ok(())
     
@@ -560,15 +498,13 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
             log.hey("");
             log.hey(&format!("{}", ">-----------------------------".dimmed()));
 
-            let mut target_uuid = None;
-
-            if let Some(uuid) = uuid {
-                target_uuid = data.iter().find(|p| p.uuid == *uuid).map(|p| p.uuid.clone());
-            }
-
-            if let Some(name) = name {
-                target_uuid = data.iter().find(|p| p.name == *name).map(|p| p.uuid.clone());
-            }
+            let target_uuid = if let Some(u) = uuid {
+                data.iter().find(|p| p.uuid == *u).map(|p| p.uuid.clone())
+            } else if let Some(n) = name {
+                data.iter().find(|p| p.name == *n).map(|p| p.uuid.clone())
+            } else {
+                None
+            };
 
 
             if let Some(uuid) = target_uuid {
@@ -594,6 +530,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                             mission: proj.mission.clone(),
                             status: proj.status.clone(),
                             is_finish: true,
+                            trace: proj.trace.clone(),
                             time_created: proj.time_created.clone()
                         };
 
@@ -602,7 +539,12 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                             *proj = my_project;
                         }
 
-                        files.write(&data, &files.mw_path.clone())?;
+                        
+                        userdata.projects_finish += 1;
+
+                        files.write(&ReturnReadType::User(userdata.clone()), &files.user_path.clone())?;
+
+                        files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
 
                         Ok(())
 
@@ -675,15 +617,13 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
             log.hey("");
             log.hey(&format!("{}", ">-----------------------------".dimmed()));
 
-            let mut target_uuid = None;
-
-            if let Some(uuid) = uuid {
-                target_uuid = data.iter().find(|p| p.uuid == *uuid).map(|p| p.uuid.clone());
-            }
-
-            if let Some(name) = name {
-                target_uuid = data.iter().find(|p| p.name == *name).map(|p| p.uuid.clone());
-            }
+            let target_uuid = if let Some(u) = uuid {
+                data.iter().find(|p| p.uuid == *u).map(|p| p.uuid.clone())
+            } else if let Some(n) = name {
+                data.iter().find(|p| p.name == *n).map(|p| p.uuid.clone())
+            } else {
+                None
+            };
 
             if *add {
 
@@ -718,7 +658,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                         log.hey_mw(&format!("{} -> {}", curr_ver, proj_version_new));
                         log.hey_mw(&format!("{} has a new version!", proj.name.bold()));
                         proj.versions.push(proj_version_new);
-                        files.write(&data, &files.mw_path.clone())?;
+                        files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
 
                         return Ok(());
 
@@ -745,7 +685,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
                         log.hey_mw(&format!("On \"{}\"'s versions", project_name));
                         log.hey_mw(&format!("{} version(s) found", project_versions.len()));
-                        log.hey(">-------------------------------------------------------------------");
+                        log.hey(">-----------------------------");
                         log.hey("");
 
                         for version in project_versions.iter() {
@@ -774,18 +714,17 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
         Commands::Yard { uuid, name, list, kill, exject } => {
 
-            log.hey_mw(&format!("{}", "On Graveyard!".red().bold()));
+            log.hey_mw(&format!("{}", "On Graveyard!".red()));
             log.hey("");
-            
-            let mut target_uuid = None;
+            log.hey(&format!("{}", ">-----------------------------".dimmed()));
 
-            if let Some(uuid) = uuid {
-                target_uuid = data.iter().find(|p| p.uuid == *uuid).map(|p| p.uuid.clone());
-            }
-
-            if let Some(name) = name {
-                target_uuid = data.iter().find(|p| p.name == *name).map(|p| p.uuid.clone());
-            }
+            let target_uuid = if let Some(u) = uuid {
+                data.iter().find(|p| p.uuid == *u).map(|p| p.uuid.clone())
+            } else if let Some(n) = name {
+                data.iter().find(|p| p.name == *n).map(|p| p.uuid.clone())
+            } else {
+                None
+            };
 
             if *kill {
 
@@ -795,9 +734,9 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
                         let proj = proj.clone();
 
-                        log.hey_mw(&format!("You are about to kill \"{}\"", proj.name));
+                        log.hey_mw(&format!("{}", "You are about to kill a project".red().italic()));
                         
-                        let want = log.quest_option("KILL? [Y/n]", vec!["Y", "n"], "n").trim().to_string();
+                        let want = log.quest_option(&format!("KILL {}? [Y/n]", proj.name.yellow().bold()), vec!["Y", "n"], "n").trim().to_string();
 
                         if want.is_empty() {
                             return Err(MyWayError::InvalidInput("Input is Empty".to_string()))
@@ -805,12 +744,17 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
                         if want == "Y" {
 
-                            log.hey_mw(&format!("{} is dead!", proj.name.red()));
+                            log.hey("");
+                            log.hey_mw(&format!("{} is {}", proj.name.yellow().bold(), "Dead!".red()));
+
                             data.retain(|p| p.uuid != &*proj.uuid);
                             graveyard.push(proj);
 
-                            files.write(&data, &files.mw_path.clone())?;
-                            files.write(&graveyard, &files.graveyard_path.clone())?;
+                            userdata.projects_killed += 1;
+
+                            files.write(&ReturnReadType::User(userdata.clone()), &files.user_path.clone())?;
+                            files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
+                            files.write(&ReturnReadType::GenericList(graveyard.to_vec()), &files.graveyard_path.clone())?;
 
                         } else {
                             log.hey_mw("Aborted. Project is still alive.");
@@ -832,7 +776,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
             if *list {
 
                 log.hey_mw(&format!("{} grave(s) found", graveyard.len()));
-                log.hey(&format!("{}", ">-------------------------------------------------------------------".dimmed()));
+                log.hey(&format!("{}", ">-----------------------------".dimmed()));
                 log.hey("");
                 
                 for proj in graveyard.iter() {
@@ -856,7 +800,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                     log.hey("");
                     log.hey_mw("Graveyard is Empty now!");
                     graveyard.clear();
-                    files.write(&graveyard, &files.graveyard_path.clone())?;
+                    files.write(&ReturnReadType::GenericList(graveyard.to_vec()), &files.graveyard_path.clone())?;
 
 
                 } else {
@@ -874,19 +818,17 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
         Commands::Reviv { uuid, name } => {
 
-            log.hey_mw(&format!("{}", "On the other S1d3!".purple()));
+            log.hey_mw(&format!("{}", "In Th3 0th3r S1d3!"));
             log.hey("");
             log.hey(&format!("{}", ">-----------------------------".dimmed()));
 
-            let mut target_uuid = None;
-
-            if let Some(uuid) = uuid {
-                target_uuid = graveyard.iter().find(|p| p.uuid == *uuid).map(|p| p.uuid.clone());
-            }
-
-            if let Some(name) = name {
-                target_uuid = graveyard.iter().find(|p| p.name == *name).map(|p| p.uuid.clone());
-            }
+            let target_uuid = if let Some(u) = uuid {
+                graveyard.iter().find(|p| p.uuid == *u).map(|p| p.uuid.clone())
+            } else if let Some(n) = name {
+                graveyard.iter().find(|p| p.name == *n).map(|p| p.uuid.clone())
+            } else {
+                None
+            };
 
 
             if let Some(uuid) = target_uuid {
@@ -909,8 +851,13 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                         graveyard.retain(|p| p.uuid != &*proj.uuid);
                         data.push(proj);
 
-                        files.write(&data, &files.mw_path.clone())?;
-                        files.write(&graveyard, &files.graveyard_path.clone())?;
+                        
+                        userdata.projects_revived += 1;
+
+                        files.write(&ReturnReadType::User(userdata.clone()), &files.user_path.clone())?;
+
+                        files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
+                        files.write(&ReturnReadType::GenericList(graveyard.to_vec()), &files.graveyard_path.clone())?;
 
                     } else {
                         log.hey_mw("Aborted. Project is still dead.");
@@ -938,15 +885,13 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
             log.hey("");
             log.hey(&format!("{}", ">-----------------------------".dimmed()));
 
-            let mut target_uuid = None;
-
-            if let Some(uuid) = uuid {
-                target_uuid = data.iter().find(|p| p.uuid == *uuid).map(|p| p.uuid.clone());
-            }
-
-            if let Some(name) = name {
-                target_uuid = data.iter().find(|p| p.name == *name).map(|p| p.uuid.clone());
-            }
+            let target_uuid = if let Some(u) = uuid {
+                data.iter().find(|p| p.uuid == *u).map(|p| p.uuid.clone())
+            } else if let Some(n) = name {
+                data.iter().find(|p| p.name == *n).map(|p| p.uuid.clone())
+            } else {
+                None
+            };
 
 
             if let Some(uuid) = target_uuid {
@@ -971,6 +916,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                             your_think: proj.your_think.clone(),
                             mission: proj.mission.clone(),
                             status: new_status,
+                            trace: proj.trace.clone(),
                             is_finish: proj.is_finish.clone(),
                             time_created: proj.time_created.clone()
                         };
@@ -980,7 +926,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
                             *proj = my_project;
                         }
 
-                        files.write(&data, &files.mw_path.clone())?;
+                        files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
 
                         Ok(())
 
@@ -1009,15 +955,13 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
             log.hey("");
             log.hey(&format!("{}", ">-----------------------------".dimmed()));
 
-            let mut target_uuid = None;
-
-            if let Some(uuid) = uuid {
-                target_uuid = data.iter().find(|p| p.uuid == *uuid).map(|p| p.uuid.clone());
-            }
-
-            if let Some(name) = name {
-                target_uuid = data.iter().find(|p| p.name == *name).map(|p| p.uuid.clone());
-            }
+            let target_uuid = if let Some(u) = uuid {
+                data.iter().find(|p| p.uuid == *u).map(|p| p.uuid.clone())
+            } else if let Some(n) = name {
+                data.iter().find(|p| p.name == *n).map(|p| p.uuid.clone())
+            } else {
+                None
+            };
 
             if let Some(uuid) = target_uuid {
 
@@ -1043,7 +987,8 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
                             log.hey_mw(&format!("{} and {} were swapped", proj_name.yellow(), other_project.name.yellow()));
                             data.swap(project_index, other_project_index);
-                            files.write(&data, &files.mw_path.clone())?;
+
+                            files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
 
                         } else {
                             return Err(MyWayError::WayLengthExceeded(format!("{} is more than {}", other_project_index, data.len())))
@@ -1059,7 +1004,8 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
                                 let it = data.remove(project_index);
                                 data.insert(0, it);
-                                files.write(&data, &files.mw_path.clone())?;
+
+                                files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
                             }
                             
                         }
@@ -1072,7 +1018,7 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
                                 let it = data.remove(project_index);
                                 data.insert(data.len(), it);
-                                files.write(&data, &files.mw_path.clone())?;
+                                files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
                             }
 
                         }
@@ -1092,6 +1038,124 @@ pub fn match_cli(command: &Commands, log: Log, files: &mut Fiman, data: &mut Pro
 
         }
 
+
+
+
+        Commands::Trace { uuid, name, list, add, remove } => {
+            
+            log.hey_mw(&format!("{}", "On Way!"));
+            log.hey("");
+            log.hey(&format!("{}", ">-----------------------------".dimmed()));
+
+            let target_uuid = if let Some(u) = uuid {
+                data.iter().find(|p| p.uuid == *u).map(|p| p.uuid.clone())
+            } else if let Some(n) = name {
+                data.iter().find(|p| p.name == *n).map(|p| p.uuid.clone())
+            } else {
+                None
+            };
+
+            if let Some(uuid) = target_uuid {
+
+                if let Some(index) = data.iter().position(|p| p.uuid == uuid) {
+
+                    if *list {
+                        for (idx, t) in data[index].trace.iter().enumerate() {
+                            log.hey_mw(&format!("{} {} {}", idx.to_string().dimmed(), "--".dimmed(), t))
+                        }
+                    }
+
+                    if *add {
+
+                        let new_trace = log.quest_mandatory("Note", "I added a new feature").trim().to_string().to_lowercase();
+
+                        let mut proj = data[index].clone();
+                        proj.trace.reverse();
+
+                        if !new_trace.is_empty() {
+                            
+                            let mut my_project: Project = Project {
+                                uuid: proj.uuid.clone(),
+                                name: proj.name.clone(),
+                                description: proj.description.clone(),
+                                stack: proj.stack.clone(),
+                                versions: proj.versions.clone(),
+                                your_think: proj.your_think.clone(),
+                                mission: proj.mission.clone(),
+                                status: new_trace.clone(),
+                                trace: proj.trace.clone(),
+                                is_finish: proj.is_finish.clone(),
+                                time_created: proj.time_created.clone()
+                            };
+
+                            my_project.trace.push(new_trace);
+
+                            if let Some(proj) = data.iter_mut().find(|p| p.uuid == my_project.uuid) {
+                                *proj = my_project;
+                            }
+
+                            let proj_name = data[index].name.clone();
+
+
+                            files.write(&ReturnReadType::GenericList(data.to_vec()), &&files.mw_path.clone())?;
+                            log.hey_mw(&format!("A New Trace was added to {}", proj_name.yellow()));
+
+                        } else {
+                            log.hey("");
+                            return Err(MyWayError::InvalidInput("Your input is empty".to_string()))
+                        }
+                    
+                    }
+
+                    if let Some(idx) = remove {
+
+                        let mut proj = data[index].clone();
+                        proj.trace.remove(*idx);
+                        let proj_name = proj.name.clone();
+
+                        data[index] = proj;
+
+                        files.write(&ReturnReadType::GenericList(data.to_vec()), &files.mw_path.clone())?;
+
+                        log.hey_mw(&format!("A note was removed from {}", proj_name.yellow()));
+
+                    }
+
+
+                } else {
+                    return Err(MyWayError::ProjectNotFound("Project not found.".into()))
+                }
+
+            } else {
+                log.hey_mw("Use '--uuid' or '--name'");
+                return Err(MyWayError::InvalidInput("No identifier provided.".into()))
+            }   
+
+            Ok(())
+
+
+        }
+
+
+        
+
+        Commands::Stats {  } => {
+
+            let h_user = serde_json::to_value(&userdata).map_err(|e| MyWayError::IoError(e.into()))?; 
+            let map : HashMap<String, serde_json::Value> = serde_json::from_value(h_user).unwrap();
+
+            log.hey_mw("Your stats:");
+            log.hey("");
+
+            for (k, v) in map.iter() {
+                log.hey_mw(&format!("> {}: {}", k.dimmed(), v.to_string().yellow()));
+            }
+
+            log.hey("");
+
+            Ok(())
+        
+        }
 
     }