+++ src/core/filemanager.rs
@@ -2,18 +2,30 @@ use std::fs;
 use std::env;
 use std::path::PathBuf;
 use std::io::{Error, ErrorKind};
-use dirs::data_local_dir;
-use tequel_rs::encrypt::TequelEncryption;
-use crate::core::{errors::MyWayError, project::ProjectList};
 
-use tequel_rs::encrypt::{ TequelEncrypt };
+use dirs::data_local_dir;
+use tequel_rs::encrypt::{ TequelEncryption, TequelEncrypt };
+
+use crate::core::project::{ GenericList, User };
+use crate::core::errors::MyWayError;
+
+
+#[derive(serde::Deserialize, serde::Serialize, Debug, Clone)]
+#[serde(untagged)]
+pub enum ReturnReadType {
+    GenericList(GenericList),
+    User(User)
+}
 
 pub struct Fiman {
     pub teq_encrypt: TequelEncrypt,
     pub user_private_key: String,
+
     pub doc_path: PathBuf,
+    
     pub mw_path: PathBuf,
-    pub graveyard_path: PathBuf
+    pub graveyard_path: PathBuf,
+    pub user_path: PathBuf
 }
 
 impl Fiman {
@@ -31,11 +43,15 @@ impl Fiman {
         
         let mw_path = doc_path.join("myway_projects.tql");
         let graveyard_path = doc_path.join("graveyard_projects.tql");
+        let user_path = doc_path.join("userdata.tql");
 
         let mut fiman = Self { 
             doc_path, 
+            
             mw_path, 
             graveyard_path, 
+            user_path,
+
             teq_encrypt: TequelEncrypt::new(), 
             user_private_key: "".to_string()
         };
@@ -67,7 +83,8 @@ impl Fiman {
     pub fn setup(&mut self) -> Result<(), MyWayError> {
             
         fs::create_dir_all(&self.doc_path).map_err(|e| { MyWayError::IoError(e) })?;
-              
+            
+        // Migrate older users to new structure
         let mut old_dir = data_local_dir().unwrap();
         old_dir.push("MyWayCli");
 
@@ -90,8 +107,18 @@ impl Fiman {
             fs::remove_file(old_path_graveyard).ok();
         }
 
+
+        // Create data files if already not exist
+
         if !self.mw_path.exists() {
             self.helper_setup_file_encrypt("[]".as_bytes(), &self.mw_path.clone())?;
+        }
+
+        if !self.user_path.exists() {
+            let user = User { projects_created: 0, projects_finish: 0, projects_killed: 0, projects_revived: 0, projects_giveup: 0 };
+            let user_data = serde_json::to_string_pretty(&user).map_err(|e| MyWayError::IoError(e.into()))?;
+
+            self.helper_setup_file_encrypt(user_data.as_bytes(), &self.user_path.clone())?;
         }
 
         if !self.graveyard_path.exists() {
@@ -102,16 +129,26 @@ impl Fiman {
 
     }
 
-    pub fn write(&mut self, data: &ProjectList, path: &PathBuf) -> Result<(), MyWayError> {
+    pub fn write(&mut self, data: &ReturnReadType, path: &PathBuf) -> Result<(), MyWayError> {
 
-        let data = serde_json::to_vec(data).map_err(|e| MyWayError::IoError(e.into()))?;
+        let json_data = serde_json::to_vec(data).map_err(|e| MyWayError::IoError(e.into()))?;
     
-        self.helper_setup_file_encrypt(&data, path)?;
+        let mut tmp_path = path.clone();
+        tmp_path.set_extension("tmp");
+
+        fs::write(&tmp_path, &json_data).map_err(MyWayError::IoError)?;
+
+        fs::rename(&tmp_path, path).map_err(|e| {
+            let _ = std::fs::remove_file(&tmp_path);
+            MyWayError::IoError(e)
+        })?;
+
+        self.helper_setup_file_encrypt(&json_data, path)?;
 
         Ok(())
     }
 
-    pub fn read(&mut self, path: &PathBuf) -> Result<ProjectList, MyWayError> {
+    pub fn read(&mut self, path: &PathBuf) -> Result<ReturnReadType, MyWayError> {
 
         let content = fs::read_to_string(&path).map_err(MyWayError::IoError)?;
         let enc_struct: TequelEncryption = serde_json::from_str(&content).map_err(|e| MyWayError::IoError(e.into()))?;
@@ -120,7 +157,7 @@ impl Fiman {
             MyWayError::IoError(Error::new(ErrorKind::Other, format!("TEQUEL FAILED: {}", e)))
         })?;
 
-        let result: ProjectList = serde_json::from_str(&decrypted).map_err(|e| {
+        let result = serde_json::from_str::<ReturnReadType>(&decrypted).map_err(|e| {
             MyWayError::IoError(Error::new(ErrorKind::Other, format!("SERDE FAILED: {}", e)))
         })?;
 
@@ -137,6 +174,7 @@ impl Fiman {
 
 
     pub fn get_machine_seed(&self) -> String {
+
         let user = env::var("USERNAME")
             .or_else(|_| env::var("USER"))
             .unwrap_or_else(|_| "unknown_user".to_string());